♻️ Move CSV-related functions to misc
